requirements.txt
flask
pillow
uwsgi